refactor: 이전 테스트 경로 이동.
